fix: API type 파일 변경 및 CategoryApi 관련 함수 이미 있는 type으로 재정의
