+++ client/src/types/api.type.ts
@@ -1,0 +1,14 @@
+export interface ApiResponse<T> {
+  status: number;
+  message: string;
+  data: T;
+}
+
+export interface ApiData<T> {
+  totalPages: number;
+  totalElements: number;
+  currentSize: number;
+  first: boolean;
+  last: boolean;
+  contents: T;
+}
