+++ client/src/utils/common.ts
@@ -1,18 +1,4 @@
 import { toast } from 'react-toastify';
-export interface ApiResponse<T> {
-  status: number;
-  message?: string;
-  data: T;
-}
-
-export interface ApiData<T> {
-  totalPages: number;
-  totalElements: number;
-  currentSize: number;
-  first: boolean;
-  last: boolean;
-  contents: T;
-}
 
 export interface ErrorResponse {
   status: string;
