+++ client/src/utils/useCategory.ts
@@ -1,37 +1,30 @@
-import { CategoryData } from '@type/data';
 import axios, { AxiosResponse } from 'axios';
-import { ApiData, ApiResponse } from './common';
+import { ApiData, ApiResponse } from '@type/api.type';
+import { CategoryDbType } from '@type/db.type';
+import { CategoryDtoType } from '@type/dto.type';
 
-interface ServerCategoryType {
-  categoryId: number;
-  categoryName: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
-interface CategoryType {
-  categoryId?: number;
-  categoryName?: string;
-}
-
-export const getCategory = async (): Promise<AxiosResponse<ApiResponse<ApiData<ServerCategoryType[]>>, any>> => {
-  return await axios.get<ApiResponse<ApiData<ServerCategoryType[]>>>('/categories');
+export const getCategory = async (): Promise<AxiosResponse<ApiResponse<ApiData<CategoryDbType[]>>, any>> => {
+  return await axios.get<ApiResponse<ApiData<CategoryDbType[]>>>('/categories');
 };
 
-export const createCategory = async (parameter: CategoryType): Promise<AxiosResponse<ApiResponse<CategoryType>>> => {
-  return await axios.post<ApiResponse<CategoryType>>('/categories', parameter);
+export const createCategory = async (
+  parameter: CategoryDtoType,
+): Promise<AxiosResponse<ApiResponse<CategoryDtoType>>> => {
+  return await axios.post<ApiResponse<CategoryDtoType>>('/categories', parameter);
 };
 
-export const updateCategory = async (parameter: CategoryType): Promise<AxiosResponse<ApiResponse<CategoryType>>> => {
-  return await axios.put<ApiResponse<CategoryType>>('/categories', parameter);
+export const updateCategory = async (
+  parameter: CategoryDtoType,
+): Promise<AxiosResponse<ApiResponse<CategoryDtoType>>> => {
+  return await axios.put<ApiResponse<CategoryDtoType>>('/categories', parameter);
 };
 
-export const deleteCategory = (id: number): (() => Promise<AxiosResponse<ApiResponse<CategoryType>, any>>) => {
-  return async () => await axios.delete<ApiResponse<CategoryType>>(`/categories/${id.toString()}`);
+export const deleteCategory = (id: number): (() => Promise<AxiosResponse<ApiResponse<CategoryDtoType>, any>>) => {
+  return async () => await axios.delete<ApiResponse<CategoryDtoType>>(`/categories/${id.toString()}`);
 };
 
-export const onSelect = (data: AxiosResponse<ApiResponse<ApiData<ServerCategoryType[]>>>): CategoryData[] => {
-  const newData = data.data.data.contents.map((elem: ServerCategoryType) => {
+export const onSelect = (data: AxiosResponse<ApiResponse<ApiData<CategoryDbType[]>>>): CategoryDtoType[] => {
+  const newData = data.data.data.contents.map((elem: CategoryDbType) => {
     return { ...elem, createdAt: new Date(elem.createdAt), updatedAt: new Date(elem.updatedAt) };
   });
   return newData;
